+++ tablet_qt/questionnairelib/quwaist.h
@@ -27,10 +27,11 @@
 
 class QuWaist : public QuMeasurement
 {
-    // Waist circumference in centimetres queestion type with imperial conversion
+    // Waist circumference in centimetres question type with imperial conversion
     Q_OBJECT
 public:
-    QuWaist(FieldRefPtr fieldref, QPointer<QuUnitSelector> unit_selector);
+    QuWaist(FieldRefPtr fieldref, QPointer<QuUnitSelector> unit_selector,
+            bool mandatory = true);
     void setUpFields();
 
 protected: